Upgrade to simpler supervision tree definition
