+++ lib/tilex/notifications/notifications.ex
@@ -9,7 +9,7 @@ defmodule Tilex.Notifications do
 
   ### Client API
 
-  def start_link() do
+  def start_link([]) do
     GenServer.start_link(__MODULE__, :ok, name: __MODULE__)
   end
