+++ lib/tilex/notifications/notifiers_supervisor.ex
@@ -1,7 +1,7 @@
 defmodule Tilex.Notifications.NotifiersSupervisor do
   use Supervisor
 
-  def start_link() do
+  def start_link([]) do
     Supervisor.start_link(__MODULE__, :ok, name: __MODULE__)
   end
