+++ lib/tilex/rate_limiter.ex
@@ -3,7 +3,7 @@ defmodule Tilex.RateLimiter do
 
   @table_name :rate_limiter_lookup
 
-  def start_link do
+  def start_link([]) do
     GenServer.start_link(__MODULE__, [], name: __MODULE__)
   end
