+++ mix.exs
@@ -26,7 +26,10 @@ defmodule Tilex.Mixfile do
   #
   # Type `mix help compile.app` for more information.
   def application do
-    [mod: {Tilex, []}, extra_applications: [:logger, :appsignal]]
+    [
+      mod: {Tilex.Application, []},
+      extra_applications: [:logger, :appsignal]
+    ]
   end
 
   # Specifies which paths to compile per environment.
